fix: add Numeric(10,2) precision and use Decimal type annotation for price columns

+++ backend/app/models/material.py
@@ -1,5 +1,6 @@
 # backend/app/models/material.py
 import uuid
+from decimal import Decimal
 from typing import Optional
 
 from sqlalchemy import CheckConstraint, ForeignKey, JSON, Numeric, Text, Uuid
@@ -25,8 +26,8 @@ class Material(Base):
     name: Mapped[str] = mapped_column(Text)
     sku: Mapped[str] = mapped_column(Text)
     thickness_options: Mapped[list] = mapped_column(JSON)  # e.g. [16, 18, 22]
-    price_per_m2: Mapped[float] = mapped_column(Numeric)
-    edgebanding_price_per_mm: Mapped[Optional[float]] = mapped_column(Numeric, nullable=True)
+    price_per_m2: Mapped[Decimal] = mapped_column(Numeric(10, 2))
+    edgebanding_price_per_mm: Mapped[Optional[Decimal]] = mapped_column(Numeric(10, 2), nullable=True)
     s3_albedo: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
     s3_normal: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
     s3_roughness: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
